printf

+++ _hexa.c
@@ -6,7 +6,7 @@
 **/
 int _hexa(va_list args)
 {
-	unsigned int n = va_arg(args, int);
+	unsigned long int n = va_arg(args, unsigned long int);
 	char arr[200];
 	int i = 0;
 	int j;
